backend/requirements.txt
click==6.6
Flask==0.11.1
get==0.0.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
post==0.0.0
public==0.0.0
query-string==0.0.0
request==0.0.0
setupfiles==0.0.0
Werkzeug==0.11.11
